Add constants and utils folder
