+++ src/utils/constants.js
@@ -1,0 +1,17 @@
+export const RenderPosition = {
+  BEFOREBEGIN: 'beforebegin',
+  AFTERBEGIN: 'afterbegin',
+  BEFORREEND: 'beforeend',
+  AFTEREND: 'afterend'
+};
+
+export const BLANK_EVENT = {
+  destination: '',
+  type: '',
+  startDate: 0,
+  finishDate: 0,
+  price: 0,
+  offers: [],
+  photos: [],
+  description: '',
+};
