+++ src/utils/createElement.js
@@ -1,0 +1,5 @@
+export const createElement = (template) => {
+  const newElement = document.createElement('div');
+  newElement.innerHTML = template;
+  return newElement.firstChild;
+};
